+++ activerecord/test/cases/adapters/sqlite3/sqlite3_adapter_test.rb
@@ -153,10 +153,6 @@ module ActiveRecord
         if in_memory_db?
           assert_equal [{ "foreign_keys" => 1 }], @conn.execute("PRAGMA foreign_keys")
           assert_equal [{ "journal_mode" => "memory" }], @conn.execute("PRAGMA journal_mode")
-          assert_equal [{ "synchronous" => 2 }], @conn.execute("PRAGMA synchronous")
-          assert_equal [{ "journal_size_limit" => 67108864 }], @conn.execute("PRAGMA journal_size_limit")
-          assert_equal [], @conn.execute("PRAGMA mmap_size")
-          assert_equal [{ "cache_size" => 2000 }], @conn.execute("PRAGMA cache_size")
         else
           with_file_connection do |conn|
             assert_equal [{ "foreign_keys" => 1 }], conn.execute("PRAGMA foreign_keys")
@@ -848,12 +844,9 @@ module ActiveRecord
           SQLite3Adapter.strict_strings_by_default = false
         end
 
-        def with_file_connection(options = {})
-          options = options.dup
-          db_config = ActiveRecord::Base.configurations.configurations.find { |config| !config.database.include?(":memory:") }
-          options[:database] ||= db_config.database
-          conn = ActiveRecord::Base.sqlite3_connection(options)
-
+        def with_file_connection
+          db_config = ActiveRecord::Base.configurations.configs_for(env_name: "arunit", name: "primary")
+          conn = Base.sqlite3_connection db_config.configuration_hash
           yield(conn)
         ensure
           conn.disconnect! if conn